feat: add route to delete a car

+++ src/controllers/abstractController.ts
@@ -55,8 +55,14 @@ abstract class AbstractController<T> {
     const { id } = req.params;
     try {
       if (!id) return res.status(412).json({ error: this.errors.requiredId });
+      if (id.length < 24) {
+        return res.status(400).json({ error: this.errors.shortId });
+      } 
       const deletedObject = await this.service.delete(id);
-      return res.status(200).json(deletedObject);
+      if (!deletedObject) {
+        return res.status(404).json({ error: this.errors.invalidId });
+      }
+      return res.status(204).json(deletedObject);
     } catch (err) {
       return res.status(500).json({ error: this.errors.internal });
     }
